move install_ssl_options_conf functionality to common
