+++ certbot-apache/certbot_apache/tests/util.py
@@ -32,10 +32,6 @@ class ApacheTest(unittest.TestCase):  # pylint: disable=too-few-public-methods
             test_dir=test_dir,
             pkg="certbot_apache.tests")
 
-        self.ssl_options = common.setup_ssl_options(
-            self.config_dir, constants.os_constant("MOD_SSL_CONF_SRC"),
-            constants.MOD_SSL_CONF_DEST)
-
         self.config_path = os.path.join(self.temp_dir, config_root)
         self.vhost_path = os.path.join(self.temp_dir, vhost_root)
